requirements.txt
fastapi>=0.110
uvicorn[standard]
openai>=1.12.0
python-dotenv
pydantic>=2.0
scikit-learn
scipy
joblib
numpy
slowapi
